LLM-Based Risk Assessment(Cloudflare is Not Available)/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
numpy
scikit-learn
joblib
pydanticflask
